feat: adjust register (#1)
